surrogate hack for fixed search page height
when the results is long enough that the page has a scroll bar, scroll
down to the bottom of the page, then mouse over the search area as to
interact with it, the page shifts upward slightly.

on mouse over, the search area becomes fixed so it no longer is included
toward the document height, so the document becomes shorter, since the
viewport is at the end, it goes up a bit and the viewport scrolls up.

with this, an invisible surrogate element switches from display none to
display block when the search area becomes fixed

it's stupid and it doesn't fix another issue where the search is 10px
too close to the element above it sometimes

i should probably give up on sticky because it still doesn't work on
firefox mobile lol https://bugzilla.mozilla.org/show_bug.cgi?id=1813213